Updated openstack/openstack Project: openstack/horizon  f26533dd445e5ba0e70716e7eb978ceaccb3e056
Fix action list icon use and spacing

The Icon used for delete actions was inconsistent with the rest
of the application. Also the lack of whitespace between icon
and text mean the rendering was quite compressed and ugly.

The alignment of the text button and drop-down button was broken
due to both having "pull-right" which has been fixed.

There is no need to have a bare <span class="fa"> here, it
serves no rendering purpose. Nor the doubling up of btn-default
and btn-danger in actions-delete-selected template.

I removed the elements of unit testing that enforced a specific
styling in the templates, as that is unnecessary and potentially
burdensome to future developers wishing to change the styling.

I removed the "btn-sm" class that was unncessary at least
in default theme under Chrome, and if sizing is actually desired
then it may be achieved through other means (a mixin or scss @extend).

Change-Id: I17ff3463aeea9ab98928124bdb0c14e82f5b04ae
Closes-Bug: 1526054