Updated openstack/openstack Project: openstack/nova  7abb4533f8829b256069a0464cf90b435e7604a3
cleanup: collapse wsgi serializer test hierarchy

There previously was a serializer / deserializer hierarchy because the
XML serializers shared some content with JSON ones, but not all. Now
that XML is out of the tree, this is an extra level of class hierarchy
for no reason.

Collapse the hierarchy, remove the 2 tests that were testing the base
classes (which were not used anywhere in functional code).

Change-Id: Iba739a4081f6dd8d6919a27e0ad61a38c33f1525