Updated openstack/openstack Project: openstack/oslo.config  47bd39a52953cbdb081de216f71ca36259dadc9f
replace reference to default value in option help text in docs

The help text for an option can include %(default)s and argparse will
include the default value when printing the help. Use the same technique
to replace that marker with a string referencing the default from the
table above the help text in the HTML output.

Change-Id: Ic16461ed50f3a8a0f30ddcb1a13dadb0f6bb0e7f