Updated openstack/openstack Project: openstack/glance  b4918e511f023f50880bc5e5d7b82f9cbf2db49a
Add functionality to define requests without body

This allows functions that do not accept bodies to define this in the
router file. As currently many requests will cause a 500 if a body is
supplied when the API request does not expect it.

This currently only affects the core parts of the v2 api, that is,
calls to v2/images and v2/schemas. It does not cover the "tasks" API
or the metadefs api as I was keeping this patch concise. As this
does not affect the behaviour if not included this makes no change to
the metadefs api behaviour.

DocImpact

Partial-Bug: 1475647

Change-Id: Ieb510e5516128078d40d39fd9b4f339ce64e10e7