Updated openstack/openstack Project: openstack/manila  360e47f6d5b5e8d04e9ebc314412e59f99841c66
Hide snapshots with no instances from listing

When a snapshot is being destroyed; for a small window of time
between the deletion of all of its instances and the snapshot ref
itself, the 'status' of the snapshot becomes 'None'. This provides
a bad user experience if a list/show query is made within that
window.

Fix this problem by filtering out snapshots that have this 'invalid'
status.

Corrected a few unit tests, to use 'status' as a manadatory
field in snapshot record.

Change-Id: I2ec83eecc9788278459846502625df7fde1fccbd
Closes-Bug: #1503390