Updated openstack/openstack Project: openstack/python-openstackclient  09be6a439a48c277af78346ce776de2df8830e98
Refactor "os availability zone list"

Refactor the "os availability zone list" command to make it a common
command instead of a compute-only command. Since availability zones
are common to compute, volume and network (new), this refactoring
allows availability zone support to be added for volume and network.

In addition to the refactor, unit and functional tests were added.

Change-Id: I63e9d41d229b21cd38e5a083493042c096d65e05
Partial-Bug: #1532945