Updated openstack/openstack Project: openstack-infra/puppet-nodepool  bb12dbb6466e1a53dd7a0c31eca58b04e6795eb8
Colocate a builder-worker with the service

This is an intermediate step before builders are completely broken out
into their own hosts and nodes.

As part of this work we split up the logging configs for the two daemons
which allows them to be colocated and write to different log files on
the same host.

Co-Authored-By: Clark Boylan <clark.boylan@gmail.com>
Change-Id: I990e96dde352fb7d01cc4b89d0f4f02de166b943