Updated openstack/openstack Project: openstack/designate  1a6a6fb9a3f4a31239061baf347daae03305a9c1
Fixes wording of exceptions for delete

Updated the wording of exceptions that prevent the deletion of
managed records. Original message indicated that user was
unable to update when the user actually requested a delete.
New message is "Managed records may not be deleted"
Both exceptions in designate/central/service.py

Change-Id: I27395999b2002374a74bddad71333f508354f37b
Closes-Bug: #1511230