Updated openstack/openstack Project: openstack-infra/project-config  97d67fb11a3a6f9a25d100a982f0ff32ca6ee635
Temporarily pin pip<8 for ceilometer tarballs

The openstack/ceilometer 2015.1.3 tag was created while the
stable/kilo branch used a deprecated pip option (--download-cache)
during sdist generation, which was subsequently removed in pip 8.0.0
and later. Temporarily pin pip to an earlier version while running
this job so that it can be retriggered without needing a new tag,
and then the ceilometer stable/kilo branch can be cleaned up to work
with latest pip afterward.

Change-Id: Ic751fe6d80d7eabb2d06de5a347ce4605e95f72c