Updated openstack/openstack Project: openstack/neutron  9aa91b12b6dd0092cff4728223d83c0aab361fbe
Rule, member updates are missed with enhanced rpc

The procedure to update security group rules and members in
firewall driver is called after update_port_filter call.
Because of this, new rules and member updates are not applied
on the port.

With this change the call to update rules and members
is moved before the port update call, giving a chance to
firewall drivers to update their rule and member caches.

Closes Bug: 1511782

Change-Id: I457e17c34b86f861f6e15de7c3adcb3f2b79d14e