Updated openstack/openstack Project: openstack/neutron  44571236ca28e3f061bdd2ed39180184539d9c4a
Fix check in _validate_ip_address() to fail on OSX

Don't check the IP address version if the data passed doesn't
have either a ':' or three '.', as we have already found an
invalid IP address.  This was failing on OSX because the
netaddr.IPAddress() call does not throw an exception, but
returns data assuming the address was IPv4.

Change-Id: Idd87e06e9cd0aa588014ef6cdc871b531a2908cb
Closes-Bug: #1534846