Updated openstack/openstack Project: openstack/python-troveclient  518d609fa5290c629357de1764855362faceb09b
Time to get rid of any last vestiges of slave_of

The controller (apischema and instance controller) have long
understood that slave_of is going away and has supported
replica_of. Getting rid of the slave_of in the client is now one part
of this final 'cleansing'. Look for a corresponding commit on the
controller side which will clean up the rest of it. That commit is
I87c8a14c06fa5df60ea3ed6b09412dc32fafcec7 and the URL for that review
is https://review.openstack.org/#/c/245845/

Change-Id: I9b905716b704ea9ec9332b48b05c49cccbd04c8b
Partial-Bug: #1461196
Depends-On: I87c8a14c06fa5df60ea3ed6b09412dc32fafcec7