Updated openstack/openstack Project: openstack/neutron  f2868434825b4cfe10bd079e95f5c18af44679d3
Make advertisement intervals for radvd configurable

Currently a global setting that is applied for all managed radvd
processes. Per-process setting could be done in the future.

For large clouds, it may be useful to increase the intervals, to reduce
multicast storms.

Co-Authored-By: Brian Haley <brian.haley@hpe.com>

DocImpact Router advertisement intervals for radvd are now configurable
Related-Bug: #1532338

Change-Id: I6cc313599f0ee12f7d51d073a22321221fca263f