Updated openstack/openstack Project: openstack/manila  60fe33ecd6b5369e439cad330395a49877d076d0
gluster*: refactor gluster_call

So far, GlusterManager.gluster_call was directly calling
into the given execution function, of which the expected
error type is ProcessExecutionError; while practically in
all use cases we wanted to raise upwards a GlusterfsException,
so all use cases were individually coercing the
ProcessExecutionError into a GlusterfsException. This produced
a huge amount of excess boilerplate code.

Here we include the coercion in the definiton of gluster_call
and clean out the excess code.

Partially implements bp gluster-code-cleanup

Change-Id: I0ad0478393df2cbb6d077363ebd6b91ceed21679