Updated openstack/openstack Project: openstack/horizon  91e1c2d2275845abbde1e5dc49bef871f1ca13b1
Only specify use of html5 locationProvider if base tag

This fixes a bug introduced by the use of the locationProvider to assume
a base tag is present.  For Launch Instance (Angular) it may be present
on a page that doesn't use Angular routing (this is OK as the modals
don't currently use such routing).

To test the fix, you should make sure that you can open and see the
Launch Instance Wizard (Angular) without any JavaScript errors.

On the flip side, you should enable an Angular panel (e.g. Images)
and ensure that it loads without error.  To fully test routing, you'll
need to pull the patch: https://review.openstack.org/#/c/173885/

Change-Id: I021d8fdb2e96eefad03182d251445d0b8ab2fa28
Closes-Bug: 1536681