Updated openstack/openstack Project: openstack/requirements  97a7114fc3cde80e4d64bb31f8b0322aa4a8ad35
Add "positional" package to g-r

Positional package allows for the emulation of the py3 arguments
needing to be either positional or key-word only. This has been
extracted from python-keystoneclient into a small stand-alone library
to de-duplicate code (it has been copied into multiple places as
the location in keystoneclient is considered private). The dep
graph of keystoneclient is also difficult to resolve this.

* Library is maintained by both jamie lennox and myself.
* This is a new library so it will not be in any distros
* See above for duplication of functionality
* This is released under Apache2

Change-Id: I99e5af7d49f0b9e2e6d0166454cca4e2456e916c