Updated openstack/openstack Project: openstack/training-guides  dff467184c921550c5151121fd0154b54f7ec544
Prepare the upstream training doc for Austin

New dates and form we're using that combines the effort with the
OpenStack mentoring program sponsored by the Women of OpenStack.
Removing the mentor wiki page link in hopes we can just the form results
in order to consolidate all the places we have mentor information.

Change-Id: I3cfc2eb6ba01e362434fedbbfaa0759f22924720