Updated openstack/openstack Project: openstack/keystonemiddleware  d661dc053adbd6088409ded27341166d45ff45a9
Group common PKI validation code - Refactor

The PKI token validation for PKI and PKIZ tokens both individually do
the cms validation and the revocation checking. There's no reason that
this needs to be in both the functions. We can split the decode process
and standardize the validation process to make the code flow easier to
understand.

As per the previous patch the tests all operate on _validate_offline
already so this is a straight refactoring.

Change-Id: I148c96688cace1f73a8fad80ed5104b1b6b871b7