Updated openstack/openstack Project: openstack/nova  a35d5a1345e1a5331860ba42c2dca5fdb27e8033
cleanup: add comments about the pre/post extension processing

In doing the cleanup of the rest of the wsgi stack I finally figured
out what really was going on here. This can't come out today because
of the legacyv2 code stack, however it can be largely removed when
that drops.

Change-Id: Ie7489b9eb1eb42ba3c4ea2d7ea973eb8d875f0ea