Updated openstack/openstack Project: openstack/nova  368e43f7a1deb2862e19c7e289cd8ad38c2cc91e
Add better help text to scheduler options (1)

The previous commit moved all the scheduler options to a central
location. This commit adds additional help text to those options.
Because of the large number of options, the improved help text will be
split into several patches; this is patch 1 of 7.

Blueprint centralize-config-options

Change-Id: I1aa6473d44f2be1c39c39e8532fffdda93a95e26