Updated openstack/openstack Project: openstack/nova  a162c766fcc51de340d6b07fc96f30d4839eda97
Add better help text to scheduler options (2)

The previous commit moved all the scheduler options to a central
location. This commit adds additional help text to those options.
Because of the large number of options, the improved help text will be
split into several patches; this is patch 2 of 7.

Blueprint centralize-config-options

Change-Id: I1ba6feaf215f62d1730340181976f25c26150b56