Updated openstack/openstack Project: openstack/keystone  ce549fda8cadba4fa99dc800175a66e763adbb53
Remove duplicate LDAP test class

test_backend_ldap is structured as a series of full test runs of
test_backend.IdentityTests in various different LDAP configurations.
With the removal of the LDAP backend for assignment, two of these
configurations (LDAPIdentity and LdapIdentitySQLAssignment) are in
fact identical.

This patch removes LdapIdentitySQLAssignment, along with minor
restructuring to ensure dependant classes still work.

There is much more than can be done to rationalize the overall
LDAP tests, but this is a quick win (given that the ldap tests
take some of longest to run of our unit tests).

Change-Id: Ifc6b58512169a7b9793f8ad5de6e027838fb7fac