Updated openstack/openstack Project: openstack/glance-specs  1f31bb0698ca88b9ae04d89ebf72762cb03780de
Prevention of Unauthorized errors in Swift driver

This change proposal introduces a new mechanism for uploading/downloading
chunked files to Swift store, that prevents possible 401 errors if user
token expires during the operation. It helps to improve the reliability
of image uploads and downloads by providing the ability to reauthenticate
during the process, thereby enhancing the stability of the entire system.

Implements bp prevention-of-401-in-swift-driver

Change-Id: Ib18b626b7545c26d2c96c94db78030a80214e18e