Updated openstack/openstack Project: openstack/openstackdocstheme  0be82d9c7366b58f438bea545889868d823335b0
Hide duplicate titles and empty tocs in generated content

At the moment we have two h1 titles, in the content generated by
Sphinx and in the titlerow.html file. This change will hide the h1 title
in the generated content.

At the moment the title of a document is displayed in the local toc. This
change will hide the title of a document in the local toc.

At the moment the local toc is also displayed when it only contains the
title of the document. This change will not display the local toc if it
conly contains the title of the document.

http://lists.openstack.org/pipermail/openstack-docs/2016-January/008180.html

Change-Id: I48b923a67010c92945c46ab00de0b090ba7cd465