Updated openstack/openstack Project: openstack-infra/project-config  b1dea8f75ebd9955a645a971000d2b958fd71fd6
puppet: continue testing for puppet-openstack-spec-helper

Continue the testing with a beaker-rspec job (non-voting) in the check
queue.

This is a follow-up of testing done in I9c7d9efae0907e4cc4c14d76dcf44ef32f78a785
but with functional testing job.
Once the job is passing, we will enable in the gate queue and enable
voting.

Change-Id: I52ddbda1b7f06db8dbdf327234e104d0eabea5d8