Updated openstack/openstack Project: openstack/tripleo-heat-templates  4c2e66a6d0dce4493f200f359dea87875e5f86aa
Allow container template to recognize an update

The deployment resource looks for a change in name when
running an update.  If there is no change in containers,
docker will recognize that and the deployment will return.
If there is a new available container, docker will swap out
the old running container for a new one.

Change-Id: I60d45b5ef45714e6e0140dfc80c14d6a12701f32