Updated openstack/openstack Project: openstack/ceilometer  d0f1db6ebce4a69c70539b8c7c818438dc8600da
Fix ceilometer floatingip pollster

The existing floatingip pollster talks to nova api
to get the floatingip data. There are limitations
in nova api wrt returning this info for all the tenants
as stated in the bug#1402514. This patch changes
the pollster to use the neutron api to get this
data instead.

Considering this is a network related pollster and in most
cases network networking manages the floating ips now,
it makes sense to get this data from neutron.

Closes-Bug: #1536338

Change-Id: I372e3a85b34f90ff9aba842d9598b468f90fe94f