Updated openstack/openstack Project: openstack/swift  c97df893f1609179f41782727fa9c7f2e4e26a3f
Use calendar.timegm not time.mktime

time.mktime produces a POSIX timestamp that takes into account the local
timezone, but we're parsing a Last-Modified header with a known (GMT)
timezone. As a result, if one proxy-server is configured with a timezone
other than GMT/UTC, we may have previous versions in the wrong order.

Change-Id: I320e2368b243f4245725e73bfabc7ad19bc5bacb