Updated openstack/openstack Project: openstack/cinder  fbeb687ea2715d40e6bc600268d9af544851216f
Disallow transferring volume in consistency group

Currently, transferring a volume in a consistency group would
success. But consistency group is not designed to span multiple
tenants. If the accepted volume is attached to an instance, and
the original user try to delete the CG, it would fail and could
not find the attachment.

This patch fixes it by adding a check in transfer API accept. It
raises InvalidVolume exception if the volume being transferred
belongs to a consistency group.

Change-Id: I597d32a301a1ded87ba711de6168995b5f62c4d8
Closes-bug: #1499584