Updated openstack/openstack Project: openstack/nova  3eb190c4cfc0eefddac6c2cc1b94a699fb1687f8
REST API changes for user settable server description

This patches adds changes to the Nova REST API to allow
users to create a server with a description, rebuild
a server with a description, update the description,
and get the description in the server details.

Note: Future commits will be done to support the server
description in python-novaclient and openstack-client.

APIImpact

Implements blueprint: user-settable-server-description

Change-Id: I74b1a340c5ab98fdea2186e87dd13f42ce7c7661