Updated openstack/openstack Project: openstack-infra/nodepool  1fa3ee77791f7c5dd4faa38a2a8012662f6e88a8
Cleanup bulder jobs and add disconnect support

If nodepool disconnects from the gear server then all outstanding jobs
are removed. Before this change we would not detect the disconnect and
would wait forever for the outstanding jobs to complete.

Moved the job tracking in to job subclasses to make the state machine
more clear.

Change-Id: Ifaf6522ba54156a2915bab9416eefa7bd14fe70b
Depends-On: If5352a373d5fa61dd1ee661f4a37976b3447dd9d