Updated openstack/openstack Project: openstack-infra/project-config  d630baaabd1b55ce598b269a792b0bd894795e0e
Fix zuul cloner macro for requirements repo

When using zuul-git-prep-upper-constraints on things like the pep8 job
for requirements, we end up producing a zuul cloner file that
references openstack/requirements twice. Zuul cloner is sad when this
happens.

As a work around, just don't added requirements. Indentation here is
important, which is why the shell looks a little wonky.

Change-Id: I3d81c8d3cf654f0e1d50b7d59856335d6b182222