Updated openstack/openstack Project: openstack/nova  d6ead5eb7bd9a5d3c6a103c6c6a5fa7f2109c092
Log retries at INFO level per guidelines

The cross-project guidelines state that positive, normal action
should be logged at INFO. Retries are a measure of how efficient
the scheduler is running, but are not "alarming". Seeing one here
and there is normal, but seeing many is cause for alarm. This should
not be buried in a debug log.

Change-Id: I530519c18c5b5acd21885fe551f3f6e39b46137a