Updated openstack/openstack Project: openstack-infra/elastic-recheck  b58d8b4cf45057c3c622947ef3d840291114217f
Add query for Ironic bug 1537076

This commit adds a query for a bug that's hitting all ironic jobs, even
on stable branches, where we aren't able to get a keystone token. This
shows up in devstacklog.txt and screen-n-cpu, but the search is easier
on the error in devstacklog so that's what I've included here.

Kibana shows 23 unique hits in 2 days, and 40 hits in the last 7 days.

Change-Id: I5a5d0970c00ab96d01e08b62ae0a64a1584b90f6
Related-bug: #1537076