Updated openstack/openstack Project: openstack/neutron  2c52dc40f04964e08659abf2d3d1a29c60b0a7fe
Neutron review tool use message instead of topics for bugs

After the upgrade of Gerrit the message field works as expected
so the review tool can use it instead of filtering using the
topic. This implies that having the bug number in the commit
message is enough for the patch to be included in the dashboard.

Change-Id: I6e61643fe862f842d731abaacfd1b18881e0c45e