Updated openstack/openstack Project: openstack/swift  f1989c5e8bcc3b5613b5bdbf7526f5da0c602830
Fix handling of "Permission Denied" error from NamedTemporaryFile function

If "Permission Denied" has happen in NamedTemporaryFile function in
dump_recon_cache method, swift will log a message of reference to a variable
without assignment and not log a message of "Permission Denied".
This patch fixes the handling and add an unit test.

Co-Authored-By: Kota Tsuyuzaki <tsuyuzaki.kota@lab.ntt.co.jp>
Change-Id: Iafdd94905e9e9c81f5966a923324b50c18fcf592