Updated openstack/openstack Project: openstack/nova  f22950816928393a4ec4eb75c6e12ad8640e310c
VMware: Handle image size correctly for OVA and streamOptimized images

The image size is different from the virtual disk size when the image is
streamOptimized or OVA. In this case we need to use the size of the
virtual disk which belongs to the temporary VM created by ImportVApp.
This works for both vSAN and VMFS datastores.

Related-Bug: #1240373
Related-Bug: #1472955

Change-Id: Ie7df9a78632ff01453b07ee9fed3dffdd8b4a0c7