Updated openstack/openstack Project: openstack/sahara-dashboard  9e32872fb36f16ff9ea4d06c832b98bb8a3de515
Fixing imports and no longer always running int tests

The horizon.utils.urlresolvers module no longer
exists, so we need to use django.core.urlresolvers.
The integration tests were always being run when
run_tests.sh was invoked.  That is not desirable.  This
patch addresses that by only running them when the
appropriate flags are passed.  Additionally, the
nose-exclude package is added to test-requriements.txt
to allow the exclusion of the integration test directories.  These 2 fixes are presented together
because tests would not pass individually, with each
patch failing on the other problem.

Change-Id: I5d3ccc8c8969992073f122f63df2c430f690a06c
Closes-Bug: #1538265
Closes-Bug: #1538159