Updated openstack/openstack Project: openstack/taskflow  f4eac0ff4e083ad02833d50f18bab250fc1d1553
Attempt to cancel active futures when suspending is underway

Instead of waiting on futures to complete that may still be
in an executors backlog (aka, not even started yet) we can
avoid even executing (and waiting on these to finish) by
forcing them to be cancelled and handling this when analyzing
the futures result (by doing nothing with the cancelled result
of a future).

This saves some time when doing suspension with executors that
have a very small pool of workers that actually run tasks/atoms.

Change-Id: Ie79466c44ac6af5ff8936f192ebce4428aa12d98